Refactor CC3Scene drawing and add framebuffer GL access.
Refactor CC3Scene drawScene method. Add drawSceneContent method for customizing
app scene drawing. Add framebuffer methods to CC3OpenGL cluster. Viewport tracked
per camera. Change CC3Scene iVar naming. Improve compilation warning resilience,
including shadow variables and pointer-interger conversion. Add additional
compilation warnings to CC3DemoMashUp project. Fix issue in resetting bound
texture in 2D cache. Fix issue with CC3GLSLVariable size tracking.

+++ Demos/CC3DemoMashUp/CC3DemoMashUp/Classes/Controls/CCNodeAdornments.m
@@ -81,7 +81,7 @@
 
 // A (hopefully) unique tag that identifies the currently activated fade-in
 // or fade-out action. This tag is used to cancel the action if needed. 
-#define kFadeActionTag 0xfade0001
+#define kFadeActionTag		(NSInteger)0xfade0001
 
 
 @implementation CCNodeAdornmentOverlayFader
